LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-alignment-style-changes-003-expected.txt


FAIL .before 1 assert_equals: 
<div id="item3" data-expected-width="30" data-expected-height="50" data-offset-y="100" class=" before" data-offset-x="20"></div>
offsetLeft expected 20 but got 0
PASS .before 2 
FAIL .before 3 assert_equals: 
<div id="item1" data-expected-width="20" data-expected-height="50" data-offset-y="0" class=" before" data-offset-x="30"></div>
offsetLeft expected 30 but got 0
